training/requirements.txt
torch>=2.2.0
torchvision>=0.17.0
timm>=1.0.3
numpy>=1.26.0
pandas>=2.1.0
scikit-learn>=1.4.0
matplotlib>=3.8.0
seaborn>=0.13.0
tqdm>=4.66.0
PyYAML>=6.0.1
Pillow>=10.0.0
onnx>=1.16.0
onnxscript>=0.1.0
